collator-protocol: route reputation timestamps through Clock
Adds a `now_ms: u128` parameter to `Backend::process_bumps` so the caller
threads the current wall-clock timestamp from the injected `Clock`
instead of each `Backend` impl reaching for `SystemTime::now()`.

`PeerManager` now holds an `Arc<dyn Clock>`; `startup()` takes one.
`Db::bump_reputations` accepts the timestamp directly. `PersistentDb`
forwards it. Existing tests pass `0` as a placeholder timestamp.

Also widens `Clock::timestamp_millis()` from `u64` to `u128` so
storage and trait line up; the single `sp_timestamp::Timestamp::new`
call site explicitly downcasts to `u64`.

Initial plan called this a `BlobStore` abstraction; the existing
`Backend` trait already isolates disk I/O, so passing a timestamp
parameter is the smaller, cleaner equivalent. (Confirmed with author.)

+++ polkadot/node/network/collator-protocol/src/clock.rs
@@ -43,7 +43,7 @@ pub trait Clock: Send + Sync {
 
 	/// Wall-clock millisecond timestamp since the UNIX epoch. Used for slot math and persistence
 	/// timestamps; not monotonic.
-	fn timestamp_millis(&self) -> u64;
+	fn timestamp_millis(&self) -> u128;
 }
 
 /// Production clock backed by `std::time` and `futures_timer`.
@@ -60,10 +60,10 @@ impl Clock for SystemClock {
 		Box::pin(futures_timer::Delay::new(dur))
 	}
 
-	fn timestamp_millis(&self) -> u64 {
+	fn timestamp_millis(&self) -> u128 {
 		SystemTime::now()
 			.duration_since(UNIX_EPOCH)
-			.map(|d| d.as_millis() as u64)
+			.map(|d| d.as_millis())
 			.unwrap_or(0)
 	}
 }
